Replace re-patching with saving mods
