+++ inject.py
@@ -1,9 +1,10 @@
 import zipfile as zf
-import os, shutil, json, keyboard, sys, pathlib
+import os, shutil, json, keyboard, sys, pathlib, utils
 from bs4 import BeautifulSoup
 from time import sleep
 
 def patch_spotify(spotify_path, delete_data = False):
+    utils.clear()
     if delete_data:
         delete_local_files()
 
@@ -30,7 +31,8 @@ def patch_spotify(spotify_path, delete_data = False):
     print("Patch applied!\nPress enter to continue...")
     keyboard.wait("Enter")
 
-def unpatch_spotify(spotify_path):
+def unpatch_spotify(spotify_path, delete_data = True):
+    utils.clear()
     extract_xpui(spotify_path)
     
     print("Undoing modifications...")
@@ -45,7 +47,7 @@ def unpatch_spotify(spotify_path):
     open("xpui-spa/index.html", "wb").write(soup.prettify("utf-8"))
 
     compile_xpui(spotify_path)
-    delete_local_files()
+    if delete_data: delete_local_files()
     clean_up()
 
     print("SpotMod has been uninstalled.\nPress enter to quit...")
@@ -89,6 +91,7 @@ def clean_up():
     os.remove("xpui.spa")
 
 def add_mod(mod_path, spotify_path):
+    utils.clear()
     mod_id = os.path.basename(mod_path)
     if os.path.exists(f"SpotMod-dat/mods/{mod_id}"):
         print("Duplicate mod detected!\nProcess aborted.")
@@ -108,6 +111,7 @@ def add_mod(mod_path, spotify_path):
     keyboard.wait("Enter")
 
 def remove_mod(mod_id, mod_ids, spotify_path):
+    utils.clear()
     print("Editing data.json...")
     data = json.load(open("SpotMod-dat/data.json"))
     data["mods"].pop(mod_ids.index(mod_id))
@@ -121,7 +125,8 @@ def remove_mod(mod_id, mod_ids, spotify_path):
     compile_xpui(spotify_path)
     clean_up()
 
-def enable_mod(mod_id, mod_ids, spotify_path, enable = True):
+def toggle_mod(mod_id, mod_ids, spotify_path, enable = True):
+    utils.clear()
     print("Editing data.json...")
     data = json.load(open("SpotMod-dat/data.json"))
     mod_data = data["mods"][mod_ids.index(mod_id)]
